[microvm] E: Rename I/O Thread Module
