+++ src/microvm/src/vmm/hyperlight/mod.rs
@@ -12,7 +12,7 @@ extern crate kvm_ioctls;
 
 use crate::{
     Gateway,
-    io::{
+    io_thread::{
         ControlCommand,
         ControlCommandResponse,
         IoThread,
